refactor(core): fold keyed lookup into lookup transform
